requirements.txt
streamlit==1.44.1
plotly==6.0.1
pandas==2.2.2
numpy==2.1.1
folium==0.19.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
haversine==2.9.0
streamlit_folium==0.25.0
pillow==10.4.0
